feat: extract comments into dedicated app
