+++ core/forms.py
@@ -10,7 +10,7 @@ import mistune
 import requests
 from PIL import Image
 
-from .models import Comment, Community
+from .models import Community
 from .utils.link_safety import check_url_safety
 
 
@@ -157,25 +157,6 @@ class PostForm(forms.Form):
             self.add_error("post_type", "Invalid post type.")
 
         return cleaned
-
-
-class CommentForm(forms.ModelForm):
-    body = forms.CharField(
-        widget=forms.Textarea(
-            attrs={"rows": 3, "data-editor": "1", "maxlength": 10000, "data-max": 10000}
-        ),
-        validators=[MaxLengthValidator(10000)],
-    )
-
-    class Meta:
-        model = Comment
-        fields = ["body"]
-
-    def clean_body(self):
-        body = (self.cleaned_data.get("body") or "").strip()
-        if not body:
-            raise forms.ValidationError("Comment cannot be empty.")
-        return body
 
 
 class CommunityCreateForm(forms.ModelForm):
