+++ core/signals.py
@@ -4,7 +4,8 @@ from django.core.cache import cache
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
-from .models import Comment, EngagementEvent
+from comments.models import Comment
+from .models import EngagementEvent
 
 
 def _cache_key(post_id):
