+++ core/tests/test_links.py
@@ -4,7 +4,8 @@ from django.template.loader import render_to_string
 from django.test import RequestFactory, TestCase
 from django.urls import reverse
 
-from core.models import Comment, Community, Post
+from core.models import Community, Post
+from comments.models import Comment
 
 
 class CommentLinkTests(TestCase):
