+++ core/tests/test_vote_auth_gate.py
@@ -3,7 +3,8 @@ from django.urls import reverse
 from django.contrib.auth import get_user_model
 from urllib.parse import quote
 
-from core.models import Community, Post, Comment
+from core.models import Community, Post
+from comments.models import Comment
 
 
 @pytest.fixture
