+++ core/urls.py
@@ -47,15 +47,6 @@ urlpatterns = [
     ),
 
     # Comments & voting
-    path("comment/<int:post_id>/reply/", core_views.comment_reply, name="comment_reply"),
-    path("comment/<int:pk>/reply-form/", core_views.comment_reply_form, name="comment_reply_form"),
-    path("comment/<int:pk>/delete/", core_views.comment_delete, name="comment_delete"),
-    path("comment/<int:pk>/edit/", core_views.comment_edit, name="comment_edit"),
-    path("comments/new", core_views.comment_new, name="comment_new"),
-    path("comments/create", core_views.comment_create, name="comment_create"),
-    path("comments/children", core_views.comment_children, name="comment_children"),
-    path("posts/vote/<int:pk>/", core_views.vote_post, name="vote_post"),
-    path("comments/vote/<int:pk>/", core_views.vote_comment, name="vote_comment"),
     # Post moderation
     path("post/<int:pk>/edit/", core_views.post_edit, name="post_edit"),
     path(
