+++ core/views.py
@@ -43,16 +43,13 @@ from django.utils.cache import patch_cache_control
 from django.contrib.contenttypes.models import ContentType
 from django_ratelimit.core import is_ratelimited
 from django.urls import reverse
-from .forms import CommentForm, PostForm, CommunityCreateForm
-from .models import Comment, Community, Post, RecoveryCode, Report
+from .forms import PostForm, CommunityCreateForm
+from .models import Community, Post, RecoveryCode, Report
+from comments.models import Comment
 from .pagination import PAGE_SIZE
 from .services.astro import compute_post_signals, compute_user_post_summary
 from .services.feed import TAB_ORDER, RANGE_MAP, feed_queryset
-from .services.votes import (
-    AlreadyVoted,
-    cast_vote_comment_once,
-    cast_vote_post_once,
-)
+from .services.votes import AlreadyVoted, cast_vote_post_once
 from . import mod
 from .http import login_required_htmx
 
@@ -897,306 +894,6 @@ def post_edit(request, pk):
     return render(request, "core/post_form.html", context)
 
 
-@login_required
-@require_POST
-def comment_reply(request, post_id):
-    """Create a new comment on a post or comment."""
-    if _is_banned(request.user):
-        return HttpResponseForbidden("Account banned")
-    post = get_object_or_404(Post.objects.filter(is_deleted=False), pk=post_id)
-    if post.is_locked:
-        return HttpResponseForbidden("Comments locked")
-    if post.slowmode:
-        rate = f"1/{post.slowmode}s"
-        if limit_or_429(request, f"slow_{post.pk}", rate):
-            return render(request, "429.html", status=429)
-    if getattr(post, 'astro_score', None) and post.astro_score.score >= settings.ASTRO_SLOWMODE_THRESHOLD:
-        if limit_or_429(request, f'astro_slow_{post.pk}', settings.ASTRO_SLOWMODE_RATE):
-            return render(request, '429.html', status=429)
-
-
-    if is_new_user(request.user):
-        if limit_or_429(request, "comment_new_user", "3/m"):
-            return render(request, "429.html", status=429)
-    else:
-        if limit_or_429(request, "comment_established", "10/m"):
-            return render(request, "429.html", status=429)
-
-
-    form = CommentForm(request.POST)
-    if not form.is_valid():
-        return HttpResponseBadRequest("Invalid comment")
-
-    parent_id = request.POST.get("parent_id")
-    parent = None
-    if parent_id:
-        parent = get_object_or_404(Comment, pk=parent_id, post=post)
-        child_seq = parent.children.count() + 1
-        path = f"{parent.path}/{child_seq:04d}"
-    else:
-        root_seq = post.comments.filter(parent__isnull=True).count() + 1
-        path = f"{root_seq:04d}"
-
-    try:
-        comment = Comment.objects.create(
-            post=post,
-            author=request.user,
-            parent=parent,
-            body=form.cleaned_data["body"],
-            path=path,
-        )
-        Post.objects.filter(pk=post.pk).update(
-            comment_count=F("comment_count") + 1
-        )
-    except (DataError, IntegrityError):
-        field, size = _find_offending_field(form)
-        logger.error("path=%s field=%s size=%s", request.path, field, size)
-        form.add_error(
-            None, "One or more fields exceed the allowed length."
-        )
-        class Dummy(SimpleNamespace):
-            def __bool__(self):
-                return False
-
-        dummy = Dummy(pk=None)
-        return render(
-            request,
-            "core/partials/comment_form.html",
-            {"form": form, "post": post, "parent": parent or dummy, "comment": dummy},
-            status=400,
-        )
-
-    if request.headers.get("HX-Request") == "true":
-        depth = comment.path.count("/")
-        return render(
-            request,
-            "core/partials/comment_row.html",
-            {"comment": comment, "depth": depth},
-        )
-
-    return redirect(
-        "post_detail",
-        community=post.community.slug,
-        pk=post.pk,
-        slug=post.slug,
-    )
-
-
-def comment_children(request):
-    """Return a batch of child comments for progressive loading."""
-
-    if request.headers.get("HX-Request") != "true":
-        return HttpResponseBadRequest("Invalid request")
-
-    parent_id = request.GET.get("parent")
-    if not parent_id:
-        return HttpResponseBadRequest("Missing parent")
-
-    try:
-        offset = int(request.GET.get("offset", 0))
-    except (TypeError, ValueError):
-        return HttpResponseBadRequest("Invalid offset")
-
-    parent = get_object_or_404(Comment, pk=parent_id)
-    children_qs = parent.children.select_related("author").order_by("path")
-    total = children_qs.count()
-    children = list(children_qs[offset : offset + PAGE_SIZE])
-    next_offset = offset + len(children)
-    remaining = max(0, total - next_offset)
-
-    html = render_to_string(
-        "core/partials/comment_children.html",
-        {
-            "children": children,
-            "parent": parent,
-            "next_offset": next_offset,
-            "remaining": remaining,
-        },
-        request=request,
-    )
-    return HttpResponse(html)
-
-
-@login_required
-@require_http_methods(["GET"])
-def comment_new(request):
-    """Return a comment form for replying via HTMX."""
-
-    if _is_banned(request.user):
-        return HttpResponseForbidden("Account banned")
-    if request.headers.get("HX-Request") != "true":
-        return HttpResponseBadRequest("Invalid request")
-
-    post_id = request.GET.get("post")
-    parent_id = request.GET.get("parent")
-    if not post_id:
-        return HttpResponseBadRequest("Missing post")
-    post = get_object_or_404(Post.objects.filter(is_deleted=False), pk=post_id)
-    if post.is_locked:
-        return HttpResponseForbidden("Comments locked")
-    parent = None
-    if parent_id:
-        parent = get_object_or_404(Comment, pk=parent_id, post=post)
-    form = CommentForm()
-    html = render_to_string(
-        "core/partials/comment_form.html",
-        {"form": form, "post": post, "parent": parent},
-        request=request,
-    )
-    return HttpResponse(html)
-
-
-@login_required
-@require_POST
-def comment_create(request):
-    """Create a new comment via HTMX."""
-
-    if _is_banned(request.user):
-        return HttpResponseForbidden("Account banned")
-    post_id = request.POST.get("post")
-    if not post_id:
-        return HttpResponseBadRequest("Missing post")
-    post = get_object_or_404(Post.objects.filter(is_deleted=False), pk=post_id)
-    if post.is_locked:
-        return HttpResponseForbidden("Comments locked")
-    if post.slowmode:
-        rate = f"1/{post.slowmode}s"
-        if limit_or_429(request, f"slow_{post.pk}", rate):
-            return render(request, "429.html", status=429)
-    if getattr(post, 'astro_score', None) and post.astro_score.score >= settings.ASTRO_SLOWMODE_THRESHOLD:
-        if limit_or_429(request, f'astro_slow_{post.pk}', settings.ASTRO_SLOWMODE_RATE):
-            return render(request, '429.html', status=429)
-    if is_new_user(request.user):
-        if limit_or_429(request, "comment_new_user", "3/m"):
-            return render(request, "429.html", status=429)
-    else:
-        if limit_or_429(request, "comment_established", "10/m"):
-            return render(request, "429.html", status=429)
-
-    form = CommentForm(request.POST)
-    if not form.is_valid():
-        return HttpResponseBadRequest("Invalid comment")
-
-    parent_id = request.POST.get("parent")
-    parent = None
-    if parent_id:
-        parent = get_object_or_404(Comment, pk=parent_id, post=post)
-        child_seq = parent.children.count() + 1
-        path = f"{parent.path}/{child_seq:04d}"
-    else:
-        root_seq = post.comments.filter(parent__isnull=True).count() + 1
-        path = f"{root_seq:04d}"
-    try:
-        comment = Comment.objects.create(
-            post=post,
-            author=request.user,
-            parent=parent,
-            body=form.cleaned_data["body"],
-            path=path,
-        )
-        Post.objects.filter(pk=post.pk).update(
-            comment_count=F("comment_count") + 1
-        )
-        post.refresh_from_db(fields=["comment_count"])
-    except (DataError, IntegrityError):
-        field, size = _find_offending_field(form)
-        logger.error("path=%s field=%s size=%s", request.path, field, size)
-        form.add_error(
-            None, "One or more fields exceed the allowed length."
-        )
-        class Dummy(SimpleNamespace):
-            def __bool__(self):
-                return False
-
-        dummy = Dummy(pk=None)
-        return render(
-            request,
-            "core/partials/comment_form.html",
-            {"form": form, "post": post, "parent": parent or dummy, "comment": dummy},
-            status=400,
-        )
-
-    if request.headers.get("HX-Request") == "true":
-        item_html = render_to_string(
-            "core/partials/comment_item.html", {"comment": comment}, request=request
-        )
-        plural = "s" if post.comment_count != 1 else ""
-        count_html = (
-            f'<a href="#comments" id="comment-count" hx-swap-oob="outerHTML">'
-            f"{post.comment_count} comment{plural}</a>"
-        )
-        return HttpResponse(item_html + count_html)
-
-    return redirect(
-        "post_detail",
-        community=post.community.slug,
-        pk=post.pk,
-        slug=post.slug,
-    )
-
-
-@require_GET
-def comment_reply_form(request, pk):
-    parent = get_object_or_404(Comment, pk=pk)
-    return render(
-        request,
-        "core/partials/reply_form.html",
-        {"parent_id": parent.pk, "post": parent.post},
-    )
-
-
-@login_required
-@require_http_methods(["GET", "POST"])
-def comment_edit(request, pk):
-    """Edit an existing comment within a 15-minute window."""
-
-    if _is_banned(request.user):
-        return HttpResponseForbidden("Account banned")
-
-    comment = get_object_or_404(Comment, pk=pk)
-    if request.user != comment.author:
-        return HttpResponseForbidden("Forbidden")
-
-    if timezone.now() - comment.created_at > timedelta(minutes=15):
-        return HttpResponseForbidden("Edit window expired")
-
-    if request.method == "GET":
-        form = CommentForm(instance=comment)
-        html = render_to_string(
-            "core/partials/comment_form.html",
-            {
-                "form": form,
-                "comment": comment,
-                "post": comment.post,
-                "parent": comment,
-            },
-            request=request,
-        )
-        return HttpResponse(html)
-
-    form = CommentForm(request.POST, instance=comment)
-    if not form.is_valid():
-        return HttpResponseBadRequest("Invalid comment")
-
-    comment.body = form.cleaned_data["body"]
-    comment.edited_at = timezone.now()
-    comment.save(update_fields=["body", "edited_at"])
-
-    if request.headers.get("HX-Request") == "true":
-        html = render_to_string(
-            "core/partials/comment_item.html", {"comment": comment}, request=request
-        )
-        return HttpResponse(html)
-
-    return redirect(
-        "post_detail",
-        community=comment.post.community.slug,
-        pk=comment.post.pk,
-        slug=comment.post.slug,
-    )
-
-
-@login_required
 @require_POST
 @csrf_protect
 def post_delete_owner(request, pk):
@@ -1308,29 +1005,6 @@ def post_domain_throttle(request, pk):
 @login_required
 @require_POST
 @csrf_protect
-def comment_delete(request, pk):
-    """Delete a comment if the requester is the author or staff."""
-    if _is_banned(request.user):
-        return HttpResponseForbidden("Account banned")
-
-    comment = get_object_or_404(Comment, pk=pk)
-    if request.user != comment.author and not request.user.is_staff:
-        return HttpResponseForbidden("Forbidden")
-
-    comment.soft_delete(request.user)
-
-    if request.headers.get("HX-Request") == "true":
-        html = render_to_string(
-            "core/partials/comment_deleted_stub.html", {"comment": comment}, request=request
-        )
-        return HttpResponse(html)
-
-    return redirect(
-        "post_detail",
-        community=comment.post.community.slug,
-        pk=comment.post.pk,
-        slug=comment.post.slug,
-    )
 
 
 @require_POST
@@ -1363,30 +1037,6 @@ def vote_post(request, pk):
 
 @require_POST
 @login_required_htmx
-def vote_comment(request, pk):
-    comment = get_object_or_404(Comment, pk=pk)
-
-    try:
-        want = int(request.POST["v"])
-    except (KeyError, TypeError, ValueError):
-        return HttpResponseBadRequest("Invalid vote")
-    if want not in (1, -1):
-        return HttpResponseBadRequest("Invalid vote")
-
-    try:
-        new_score = cast_vote_comment_once(request.user, comment, want)
-    except AlreadyVoted:
-        return HttpResponse(status=409)
-
-    if request.headers.get("HX-Request") == "true":
-        comment.refresh_from_db()
-        return render(
-            request,
-            "core/partials/vote_widget.html",
-            {"comment": comment, "voted": True},
-        )
-
-    return HttpResponse(f"<span id='comment-{comment.pk}-score'>{new_score}</span>")
 
 
 @login_required
