+++ templates/comments/comment_deleted_stub.html
@@ -5,7 +5,7 @@
   {% if comment.depth < 5 %}
   <div class="children">
     {% for child in comment.children.all %}
-      {% include "core/partials/comment_item.html" with comment=child %}
+      {% include "comments/comment_item.html" with comment=child %}
     {% endfor %}
   </div>
   {% endif %}
